2024年11月17日 22:26

+++ Test_Learning/DGUT_OJ/新生赛/2.c
@@ -6,9 +6,9 @@ int main () {
     int n;
     long long min;
     scanf("%d", &n);
-    long long* arr1 = (long long*)malloc(1000000000000LL * sizeof(long long));
+    int* arr1 = (int*)malloc(1000000000000 * sizeof(int));
     long long* arr = (long long*)malloc(n * sizeof(long long));
-    memset(arr1, 0, 1000000000000LL * sizeof(long long));
+    memset(arr1, 0, 1000000000000 * sizeof(int));
     for (int i = 0; i < n; i++) {
         scanf("%lld", (arr + i));
         *(arr1 + *(arr + i)) += 1;
@@ -18,7 +18,7 @@ int main () {
         if (min > *(arr + i)) {
             min = *(arr + i);
         }
-        printf("%lld %lld\n", min, *(arr1 + min));
+        printf("%lld %d\n", min, *(arr1 + min));
     }
     free(arr);
     free(arr1);
